plgx-esp/requirements/prod.txt
alembic==1.0.8            # via flask-migrate
amqp==2.4
pyamqp==0.0.8.5
apscheduler==3.6.0
anyjson==0.3.3

bcrypt==3.1.6
beautifulsoup4==4.7.1
billiard==3.5.0.5          # via celery
blinker==1.4

celery==4.1.1

cffi==1.12.2              # via bcrypt
contextlib2==0.5.5
cssmin==0.2.0

enum34==1.1.6
eventlet==0.24.1

flask==0.12.2
flask-bcrypt==0.7.1
flask-ldap3-login==0.9.16
flask-login==0.4.1
flask-migrate==2.4.0
flask-paginate==0.5.2
flask-script==2.0.6
flask_sockets==0.2.1
flask_socketio==3.3.2
flask-wtf==0.14.2

Flask-Assets==0.12
Flask-Caching==1.7.1
Flask-DebugToolbar==0.10.1
Flask-Mail==0.9.1
Flask-SSLify==0.1.5
Flask-SQLAlchemy==2.3.2

flower==0.9.3

gevent-websocket==0.10.1
gunicorn==19.9.0
itsdangerous==1.1.0        # via flask

jinja2==2.10             # via flask
jsmin==2.2.2

kombu==4.2.1          # via celery

ldap3==2.6.0

mako==1.0.8               # via alembic
markupsafe==1.1.1

oauthlib==3.0.1

pika==0.13.0
psycopg2-binary==2.8
pyamqp==0.0.8.5
pyasn1==0.4.5
pycparser==2.19           # via cffi
python-editor==1.0.4      # via alembic
pytz==2018.9              # via celery, flask-restful

raven==6.10.0
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0

scales==1.0.9
socketIO-client3==0.8.0
sqlalchemy-datatables==2.0.1
SQLAlchemy==1.3.10

tornado==6.0.2

unicodecsv==0.14.1

webassets==0.12.1
werkzeug==0.15.2          # via flask
wtforms==2.1              # via flask-wtf


# threat intel libs
cymon==0.1
shodan==1.11.1
virustotal-api==1.1.10
avcaesar_api==1.0.4
pysafebrowsing==0.1.1
cybercure==0.4.3
OTXv2==1.3
